Refactor PropPane Groups, Clean up imports
