+++ src/webparts/drilldown7/components/HelpPanel/Basics.tsx
@@ -1,68 +1,68 @@
-import * as React from 'react';
+// import * as React from 'react';
 
-import { Icon } from 'office-ui-fabric-react';
+// import { IHelpTable, } from '../../fpsReferences';
+// import { repoLink, } from '../../fpsReferences';
+// import { convertIssuesMarkdownStringToSpan } from '../../fpsReferences';
 
-import * as links from '@mikezimm/npmfunctions/dist/Links/LinksRepos';
+// import { createAboutRow } from '../../fpsReferences';
 
-import { IRepoLinks } from '@mikezimm/npmfunctions/dist/Links/CreateLinks';
-
-export function basicsContent( repoLinks: IRepoLinks ) {
+export function basicsContent(  ) {
 
     return null;
 
-    let messageRows = [];
+//     let messageRows = [];
 
-    let iconSize = 'large';
+//     let iconSize = 'large';
 
-    let iconStyles: any = { root: {
-        fontSize: 'x-large',
-        fontWeight: 600,
-        paddingRight: '10px',
-        paddingLeft: '10px',
-    }};
+//     let iconStyles: any = { root: {
+//         fontSize: 'x-large',
+//         fontWeight: 600,
+//         paddingRight: '10px',
+//         paddingLeft: '10px',
+//     }};
 
-    let gridIcon = <Icon iconName={"GridViewSmall"}  style={ iconStyles } />; 
-    let tilesIcon = <Icon iconName={"Tiles"}  style={ iconStyles } />; 
-    let filterIcon = <Icon iconName={"ClearFilter"}  style={ iconStyles } />;
-    let dashIcon = <Icon iconName={"ChromeMinimize"}  style={ iconStyles } />;
+//     let gridIcon = <Icon iconName={"GridViewSmall"}  style={ iconStyles } />; 
+//     let tilesIcon = <Icon iconName={"Tiles"}  style={ iconStyles } />; 
+//     let filterIcon = <Icon iconName={"ClearFilter"}  style={ iconStyles } />;
+//     let dashIcon = <Icon iconName={"ChromeMinimize"}  style={ iconStyles } />;
 
-    messageRows.push( <tr><td>Alt-Click</td><td style={{whiteSpace: 'nowrap'}}><b>HTML File Link</b></td><td>Open the html file <b>in Edit Mode</b></td></tr> );
+//     messageRows.push( <tr><td>Alt-Click</td><td style={{whiteSpace: 'nowrap'}}><b>HTML File Link</b></td><td>Open the html file <b>in Edit Mode</b></td></tr> );
 
-    messageRows.push( <tr><td></td><td></td><td></td></tr> );
+//     messageRows.push( <tr><td></td><td></td><td></td></tr> );
 
-    let thisTable = <div style={{ paddingTop: 15 }}>
-        <h2>Easter Eggs</h2>
-        <table className={ 'single-page-info-table' } style={{ width: '100%' }}>
-            <tr><th style={{ minWidth: '70px' }}>Do this...</th><th>Where</th><th>What it does</th></tr>
-            { messageRows }
-        </table>
-    </div>;
-
-
-    let thisPage = null;
-    thisPage = <div className={ null }>
-
-    { thisTable }
-  </div>;
+//     let thisTable = <div style={{ paddingTop: 15 }}>
+//         <h2>Easter Eggs</h2>
+//         <table className={ 'single-page-info-table' } style={{ width: '100%' }}>
+//             <tr><th style={{ minWidth: '70px' }}>Do this...</th><th>Where</th><th>What it does</th></tr>
+//             { messageRows }
+//         </table>
+//     </div>;
 
 
-/***
-*              d8888b. d88888b d888888b db    db d8888b. d8b   db 
-*              88  `8D 88'     `~~88~~' 88    88 88  `8D 888o  88 
-*              88oobY' 88ooooo    88    88    88 88oobY' 88V8o 88 
-*              88`8b   88~~~~~    88    88    88 88`8b   88 V8o88 
-*              88 `88. 88.        88    88b  d88 88 `88. 88  V888 
-*              88   YD Y88888P    YP    ~Y8888P' 88   YD VP   V8P 
-*                                                                 
-*                                                                 
-*/
+//     let thisPage = null;
+//     thisPage = <div className={ null }>
 
-let html1 =
-<div className={ 'single-page-info-pane' }>
-    { thisPage }
-</div>;
+//     { thisTable }
+//   </div>;
 
-return { html1: html1 };
+
+// /***
+// *              d8888b. d88888b d888888b db    db d8888b. d8b   db 
+// *              88  `8D 88'     `~~88~~' 88    88 88  `8D 888o  88 
+// *              88oobY' 88ooooo    88    88    88 88oobY' 88V8o 88 
+// *              88`8b   88~~~~~    88    88    88 88`8b   88 V8o88 
+// *              88 `88. 88.        88    88b  d88 88 `88. 88  V888 
+// *              88   YD Y88888P    YP    ~Y8888P' 88   YD VP   V8P 
+// *                                                                 
+// *                                                                 
+// */
+
+// let html1 =
+// <div className={ 'single-page-info-pane' }>
+//     { thisPage }
+// </div>;
+
+// return { html1: html1 };
 }
