+++ src/webparts/drilldown7/components/HelpPanel/FuturePlans.tsx
@@ -1,18 +1,20 @@
-import * as React from 'react';
+// import * as React from 'react';
 
-import { IRepoLinks } from '@mikezimm/npmfunctions/dist/Links/CreateLinks';
+// import { IHelpTable, } from '../../fpsReferences';
+// import { repoLink, } from '../../fpsReferences';
+// import { convertIssuesMarkdownStringToSpan } from '../../fpsReferences';
 
-import { convertIssuesMarkdownStringToSpan } from '@mikezimm/npmfunctions/dist/Elements/Markdown';
+// import { createAboutRow } from '../../fpsReferences';
 
-export function futureContent( repoLinks: IRepoLinks ) {
+export function futureContent( ) {
 
     return null;
 
-    let html1 = <div>
-        <h2>Were thinking of making this an extension so it doesn't need to be added to a page!</h2>
-    </div>;
+    // let html1 = <div>
+    //     <h2>Were thinking of making this an extension so it doesn't need to be added to a page!</h2>
+    // </div>;
 
-    return { html1: html1 };
+    // return { html1: html1 };
 
 }
