+++ src/webparts/drilldown7/components/createButtons/IconButton.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { IconButton, IIconProps, IContextualMenuProps, Stack, Link } from 'office-ui-fabric-react';
+import { IconButton, IIconProps, } from 'office-ui-fabric-react';
 
 const emojiIcon: IIconProps = { iconName: 'BarChartVerticalFill' };
